requirements.txt
click
requests
bs4
python-dotenv
setuptools